Switch to figment based configuration
The old config setup (generously setup by a contributor) was very WET and becoming a real pain to maintain.

Also took a look at confique, but that cannot handle nested, Vec configs as we need for our calendar sources. It's a shame because it can automatically generate an exmaple config file, something that will undoubtedly be needed.